Add files via upload
In this code, we first split the data into training and testing sets using the train_test_split function. Then, we use the TfidfVectorizer to convert the text data into numerical features using TF-IDF vectorization. Next, we create an instance of the LinearSVC class, which represents the Linear SVM classifier. We fit the classifier to the training data using the fit method. After that, we use the trained model to make predictions on the test set using the predict method. Finally, we evaluate the accuracy of the model by comparing the predicted labels with the true labels of the test set..